arreglado problema con la carga del dashboard principal y udgraded del modulo de crm con clietes y proveedores

+++ ventiq_admin_app/lib/models/crm/crm_metrics.dart
@@ -5,7 +5,7 @@ class CRMMetrics {
   final int activeCustomers;
   final int vipCustomers;
   final double averageCustomerValue;
-  final int loyaltyPoints;
+  final double loyaltyPoints;
 
   // Métricas de Proveedores
   final int totalSuppliers;
@@ -111,7 +111,7 @@ class CRMMetrics {
     int? activeCustomers,
     int? vipCustomers,
     double? averageCustomerValue,
-    int? loyaltyPoints,
+    double? loyaltyPoints,
     int? totalSuppliers,
     int? activeSuppliers,
     double? averageLeadTime,
